requirements.txt
streamlit>=1.32
pandas>=2.0
numpy>=1.24
openpyxl>=3.1
python-dotenv>=1.0
altair>=5.0